+++ OrderTrackingApp/OrderTrackingApp/ViewModels/OrderHistoryViewModel.cs
@@ -1,5 +1,6 @@
 ﻿using OrderTrackingApp.Models;
 using OrderTrackingApp.Objects;
+using OrderTrackingApp.Resx;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -79,17 +80,22 @@ namespace OrderTrackingApp.ViewModels
 
         private void Initialize()
         {
-            DeleteItemButtonClick = new Command<int>((id) =>
+            DeleteItemButtonClick = new Command<int>(async (id) =>
             {
-                Order item = orders.First(p => p.Id == id);
-                DAL.DAL.DeleteOrder(item);
-                RefreshItems();
+                var confirmed = await Application.Current.MainPage.DisplayAlert(AppResources.ConfirmTtile, AppResources.DeleteConfirmationMsg, AppResources.Yes, AppResources.No);
+                if (confirmed)
+                {
+                    Order item = orders.First(p => p.Id == id);
+                    DAL.DAL.DeleteOrder(item);
+                    RefreshItems();
+                }
             });
             PayItemButtonClick = new Command<int>((id) =>
             {
                 Order item = orders.First(p => p.Id == id);
                 DAL.DAL.SetOrderStatus(item, true);
                 RefreshItems();
+                MessagingCenter.Send(this, "OrderPayed");
             });
             ShowOpenButtonClick = new Command(() =>
             {
